Terrain_Recognition_Project/TerrainRecognitionSIH-main/requirements.txt
Flask==2.2.2
Flask_Cors==4.0.0
numpy==1.24.3
Pillow==9.4.0
Pillow==10.1.0
tensorflow==2.14.0
tensorflow_macos==2.14.0
gunicorn==20.0.4
Werkzeug==0.16.0
Jinja2==2.10.1
itsdangerous==1.1.0
MarkupSafe==1.1.1